Taking more advantage of LESS capabilities
The main change is the consolidation of colour information in
variables, and the use of less functions to modify a few default values
instead of having the same colour information repeated.

Also combined handheld and print within a single .less file to improve
the behaviour of less.js on mobile devices.  Combining in this way
makes it work properly on small-screen iOS, which used fullscreen
rendering instead of the style given in handheld.css (or a temporarily
existing handheld.less) when it was a separate file.  It still appears
to fail in Opera Mobile, which does not appear to apply CSS styling.